refactor(emit): split core helper types
